+++ src/components/Map/index.tsx
@@ -17,8 +17,8 @@ import { useQuery } from '@tanstack/react-query';
 import MapButton from './MapButton';
 import AirPollutionLevels from '@/components/Map/AirPollutionLevels';
 import DustState from '@/components/Dust/DustState';
-import { getAllLocation } from '@/api/location';
-import { getSidoAirQualities, getCityAirQualities } from '@/api/airQuality';
+import { getAllLocation } from '@/apis/location';
+import { getSidoAirQualities, getCityAirQualities } from '@/apis/airQuality';
 import { getDustScaleColor } from '@/utils/map';
 import { FINE_DUST, ULTRA_FINE_DUST } from '@/utils/constants';
 import useMap from '@/hooks/useMap';
